move GamePage to folder and add CSS Module
