+++ src/pages/Game/GamePage.module.css
@@ -1,0 +1,67 @@
+.layout {
+  display: flex;
+  flex-direction: column;
+  gap: 20px;
+}
+
+.info {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  gap: 20px;
+  flex-wrap: wrap;
+}
+
+.questionCard {
+  min-height: 300px;
+  display: flex;
+  flex-direction: column;
+}
+
+.actions {
+  text-align: center;
+}
+
+.timer {
+  background: rgba(255, 255, 255, 0.1);
+  padding: 10px 20px;
+  border-radius: 15px;
+  color: white;
+  font-weight: bold;
+  font-size: 1.1rem;
+  backdrop-filter: blur(10px);
+  border: 1px solid rgba(255, 255, 255, 0.2);
+}
+
+.timerWarning {
+  background: #ff5252 !important;
+  color: white !important;
+  animation: pulse 1s infinite;
+}
+
+@keyframes pulse {
+  0% { transform: scale(1); }
+  50% { transform: scale(1.05); }
+  100% { transform: scale(1); }
+}
+
+.debug {
+  margin-top: 20px;
+  padding: 10px;
+  background: #f5f5f5;
+  border-radius: 5px;
+  font-size: 12px;
+  color: #666;
+}
+
+@media (max-width: 768px) {
+  .info {
+    flex-direction: column;
+    text-align: center;
+    gap: 15px;
+  }
+  
+  .questionCard {
+    min-height: 250px;
+  }
+}
